# api/server/sdk/api/api.swagger.json
{
  "components": {
    "schemas": {
...
      "apiAlert": {
        "properties": {
          "alert_type": {
            "format": "int64",
            "title": "AlertType user defined alert type",
            "type": "string"
          },
          "cleared": {
            "format": "boolean",
            "title": "Cleared Flag",
            "type": "boolean"
          },
          "count": {
            "description": "Count of such alerts raised so far.",
            "format": "int64",
            "type": "string"
          },
          "first_seen": {
            "description": "Timestamp when such alert was raised the very first time.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "title": "Id for Alert",
            "type": "string"
          },
          "message": {
            "title": "Message describing the Alert",
            "type": "string"
          },
          "resource": {
            "$ref": "#/components/schemas/apiResourceType"
          },
          "resource_id": {
            "title": "ResourceId where Alert occured",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/apiSeverityType"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp when Alert occured",
            "type": "string"
          },
          "ttl": {
            "format": "uint64",
            "title": "Time-to-live in seconds for this Alert",
            "type": "string"
          },
          "unique_tag": {
            "title": "UniqueTag helps identify a unique alert for a given resouce",
            "type": "string"
          }
        },
        "title": "Alert is a structure that represents an alert object",
        "type": "object"
      },
...
      "apiSdkAlertsEnumerateWithFiltersRequest": {
        "description": "SdkAlertsEnumerateRequest is a request message to enumerate alerts.",
        "properties": {
          "queries": {
            "description": "It is a list of queries to find matching alerts.\nOutput of each of these queries is added to a global pool\nand returned as output of an RPC call.\nIn that sense alerts are fetched if they match any of the\nqueries.",
            "items": {
              "$ref": "#/components/schemas/apiSdkAlertsQuery"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "apiSdkAlertsEnumerateWithFiltersResponse": {
        "description": "SdkAlertsEnumerateResponse is a list of alerts.",
        "properties": {
          "alerts": {
            "description": "Response contains a list of alerts.",
            "items": {
              "$ref": "#/components/schemas/apiAlert"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "apiSdkAlertsQuery": {
        "description": "SdkAlertsQuery is one of the query types and a list of options.\nEach query object is one of the three query types and a list of\noptions.",
        "properties": {
          "alert_type_query": {
            "$ref": "#/components/schemas/apiSdkAlertsAlertTypeQuery"
          },
          "opts": {
            "description": "Opts is a list of options associated with one of the queries.",
            "items": {
              "$ref": "#/components/schemas/apiSdkAlertsOption"
            },
            "type": "array"
          },
          "resource_id_query": {
            "$ref": "#/components/schemas/apiSdkAlertsResourceIdQuery"
          },
          "resource_type_query": {
            "$ref": "#/components/schemas/apiSdkAlertsResourceTypeQuery"
          }
        },
        "type": "object"
      },
...
      "apiSdkAlertsResourceTypeQuery": {
        "description": "SdkAlertsResourceTypeQuery queries for alerts using only resource id.",
        "properties": {
          "resource_type": {
            "$ref": "#/components/schemas/apiResourceType"
          }
        },
        "type": "object"
      },
...
      "apiSdkCloudBackupCatalogRequest": {
        "properties": {
          "backup_id": {
            "title": "Id of the backup",
            "type": "string"
          },
          "credential_id": {
            "title": "Credential id describe the credentials for the cloud",
            "type": "string"
          }
        },
        "title": "Defines a request to get catalog of a backup stored by a cloud provider",
        "type": "object"
      },
      "apiSdkCloudBackupCatalogResponse": {
        "properties": {
          "contents": {
            "items": {
              "type": "string"
            },
            "title": "Contents is listing of backup contents",
            "type": "array"
          }
        },
        "title": "Defines a response containing the contents of a backup stored by a cloud provider",
        "type": "object"
      },
...
      "apiSdkCloudBackupEnumerateWithFiltersRequest": {
        "description": "* For a specific volume in current cluster: Set `src_volume_id` to your desired volume id\nand do not provide `cluster_id` and `all`.\n* For a specific volume in a specific cluster: Set `src_volume_id` to your desired volume id\nand specify `cluster_id`.\n* For a specific volume in all clusters: Set `src_volume_id` to your desired volume id\nand set `all` to true, do not provide `cluster_id`.\n* For all volumes in current cluster: do not provide `cluster_id`, `volume_id` and `all`.\n* For all volumes in a specific cluster: Set `cluster_id` to your desired cluster id\nand do not provide `volume_id` and `all`.\n* For all volumes in all clusters: Set `all` to true do not provide `volume_id` and `cluster_id`.",
        "properties": {
          "all": {
            "description": "(optional) All indicates if the request should show cloud backups for all clusters or the current cluster.",
            "format": "boolean",
            "type": "boolean"
          },
          "cluster_id": {
            "title": "(optional) Cluster id specifies the cluster for the request",
            "type": "string"
          },
          "credential_id": {
            "title": "Credential id is the credential for cloud to be used for the request",
            "type": "string"
          },
          "src_volume_id": {
            "description": "(optional) Source id of the volume for the request.",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/apiSdkCloudBackupStatusType"
          }
        },
        "title": "Defines a request to list the backups stored by a cloud provider.\nThe following combinations can be used to get cloud backup information:",
        "type": "object"
      },
      "apiSdkCloudBackupEnumerateWithFiltersResponse": {
        "properties": {
          "backups": {
            "items": {
              "$ref": "#/components/schemas/apiSdkCloudBackupInfo"
            },
            "type": "array"
          }
        },
        "title": "Defines a response which lists all the backups stored by a cloud provider",
        "type": "object"
      },
      "apiSdkCloudBackupHistoryItem": {
        "properties": {
          "src_volume_id": {
            "title": "SrcVolumeID is volume ID which was backedup",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/apiSdkCloudBackupStatusType"
          },
          "timestamp": {
            "format": "date-time",
            "title": "TimeStamp is the time at which either backup completed/failed",
            "type": "string"
          }
        },
        "title": "SdkCloudBackupHistoryItem contains information about a backup for a\nspecific volume",
        "type": "object"
      },
      "apiSdkCloudBackupHistoryResponse": {
        "properties": {
          "history_list": {
            "items": {
              "$ref": "#/components/schemas/apiSdkCloudBackupHistoryItem"
            },
            "title": "HistoryList is list of past backups on this volume",
            "type": "array"
          }
        },
        "title": "Defines a response containing a list of history of backups to a cloud provider",
        "type": "object"
      },
      "apiSdkCloudBackupInfo": {
        "properties": {
          "id": {
            "title": "This is the id as represented by the cloud provider",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Metadata associated with the backup",
            "type": "object"
          },
          "src_volume_id": {
            "title": "Source volumeID of the backup",
            "type": "string"
          },
          "src_volume_name": {
            "title": "Name of the sourceVolume of the backup",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/apiSdkCloudBackupStatusType"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp is the timestamp at which the source volume\nwas backed up to cloud",
            "type": "string"
          }
        },
        "title": "SdkCloudBackupInfo has information about a backup stored by a cloud provider",
        "type": "object"
      },
...
      "apiSdkCloudBackupRequestedState": {
        "default": "SdkCloudBackupRequestedStateUnknown",
        "description": "- SdkCloudBackupRequestedStateUnknown: Unknown state\n - SdkCloudBackupRequestedStatePause: Pause the backup or restore\n - SdkCloudBackupRequestedStateResume: Resume the backup or restore\n - SdkCloudBackupRequestedStateStop: Stop a backup or restore",
        "enum": [
          "SdkCloudBackupRequestedStateUnknown",
          "SdkCloudBackupRequestedStatePause",
          "SdkCloudBackupRequestedStateResume",
          "SdkCloudBackupRequestedStateStop"
        ],
        "title": "SdkCloudBackupRequestedState defines states to set a specified backup or restore\nto or from a cloud provider",
        "type": "string"
      },
...
      "apiSdkCloudBackupSchedCreateRequest": {
        "properties": {
          "cloud_sched_info": {
            "$ref": "#/components/schemas/apiSdkCloudBackupScheduleInfo"
          }
        },
        "title": "Defines a request to create a schedule for volume backups to a\ncloud provider",
        "type": "object"
      },
      "apiSdkCloudBackupSchedCreateResponse": {
        "properties": {
          "backup_schedule_id": {
            "title": "Id of newly created backup schedule",
            "type": "string"
          }
        },
        "title": "Defines a response containing the id of a schedule for a volume backup\nto a cloud provider",
        "type": "object"
      },
...
      "apiSdkCloudBackupSchedEnumerateResponse": {
        "properties": {
          "cloud_sched_list": {
            "additionalProperties": {
              "$ref": "#/components/schemas/apiSdkCloudBackupScheduleInfo"
            },
            "title": "Returns list of backup schedules",
            "type": "object"
          }
        },
        "title": "Defines a response containing a map listing the schedules for volume\nbackups to a cloud provider",
        "type": "object"
      },
...
      "apiSdkCloudBackupStateChangeRequest": {
        "properties": {
          "requested_state": {
            "$ref": "#/components/schemas/apiSdkCloudBackupRequestedState"
          },
          "task_id": {
            "title": "Describes the backup/restore task\nstate change is being requested",
            "type": "string"
          }
        },
        "title": "Defines a request to change the state of a backup or restore to or\nfrom a cloud provider",
        "type": "object"
      },
      "apiSdkCloudBackupStateChangeResponse": {
        "title": "Empty response",
        "type": "object"
      },
...
      "apiSdkCloudBackupStatusType": {
        "default": "SdkCloudBackupStatusTypeUnknown",
        "description": "- SdkCloudBackupStatusTypeUnknown: Unkonwn\n - SdkCloudBackupStatusTypeNotStarted: Not started\n - SdkCloudBackupStatusTypeDone: Done\n - SdkCloudBackupStatusTypeAborted: Aborted\n - SdkCloudBackupStatusTypePaused: Paused\n - SdkCloudBackupStatusTypeStopped: Stopped\n - SdkCloudBackupStatusTypeActive: Active\n - SdkCloudBackupStatusTypeFailed: Failed\n - SdkCloudBackupStatusTypeQueued: Queued\n - SdkCloudBackupStatusTypeInvalid: Invalid",
        "enum": [
          "SdkCloudBackupStatusTypeUnknown",
          "SdkCloudBackupStatusTypeNotStarted",
          "SdkCloudBackupStatusTypeDone",
          "SdkCloudBackupStatusTypeAborted",
          "SdkCloudBackupStatusTypePaused",
          "SdkCloudBackupStatusTypeStopped",
          "SdkCloudBackupStatusTypeActive",
          "SdkCloudBackupStatusTypeFailed",
          "SdkCloudBackupStatusTypeQueued",
          "SdkCloudBackupStatusTypeInvalid"
        ],
        "title": "CloudBackup status types",
        "type": "string"
      },
...
      "apiSdkVolumeEnumerateWithFiltersRequest": {
        "properties": {
          "group": {
            "$ref": "#/components/schemas/apiGroup"
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "(optional) Labels to search",
            "type": "object"
          },
          "name": {
            "title": "(optional) Name to search",
            "type": "string"
          },
          "ownership": {
            "$ref": "#/components/schemas/apiOwnership"
          }
        },
        "title": "Defines a request to list volumes",
        "type": "object"
      },
      "apiSdkVolumeEnumerateWithFiltersResponse": {
        "properties": {
          "volume_ids": {
            "items": {
              "type": "string"
            },
            "title": "List of volumes matching label",
            "type": "array"
          }
        },
        "title": "Defines the response when listing volumes",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "OpenStorage SDK",
    "version": "0.52.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/alerts/filters": {
      "post": {
        "description": "EnumerateWithFilters allows 3 different types of queries as defined below:\n\n* Query that takes only resource type as input\n* Query that takes resource type and alert type as input and\n* Query that takes resource id, alert type and resource type as input.\n\n#### Input\nSdkAlertsEnumerateRequest takes a list of such queries and the returned\noutput is a collective ouput from each of these queries. In that sense,\nthe filtering of these queries has a behavior of OR operation.\nEach query also has a list of optional options. These options allow\nnarrowing down the scope of alerts search. These options have a\nbehavior of an AND operation.\n\n#### Examples\nTo search by a resource type in a given time window would require\ninitializing SdkAlertsResourceTypeQuery query and pass in\nSdkAlertsTimeSpan option into SdkAlertsQuery struct and finally\npacking any other such queries into SdkAlertsEnumerateRequest object.\nAlternatively, to search by both resource type and alert type, use\nSdkAlertsAlertTypeQuery as query builder.\nFinally to search all alerts of a given resource type and some\nalerts of another resource type but with specific alert type,\nuse two queries, first initialized with SdkAlertsResourceTypeQuery\nand second initialized with SdkAlertsAlertTypeQuery and both\neventually packed as list in SdkAlertsEnumerateRequest.",
        "operationId": "EnumerateWithFilters",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/apiSdkAlertsEnumerateWithFiltersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiSdkAlertsEnumerateWithFiltersResponse"
                }
              }
            },
            "description": "(streaming responses)"
          }
        },
        "summary": "Allows querying alerts.",
        "tags": [
          "OpenStorageAlerts"
        ]
      }
    },
...
    "/v1/cloudbackups/catalog": {
      "post": {
        "operationId": "Catalog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/apiSdkCloudBackupCatalogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiSdkCloudBackupCatalogResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Catalog returns a list of the contents in the backup",
        "tags": [
          "OpenStorageCloudBackup"
        ]
      }
    },
...
    "/v1/cloudbackups/enumerate/filters": {
      "post": {
        "operationId": "EnumerateWithFilters",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/apiSdkCloudBackupEnumerateWithFiltersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiSdkCloudBackupEnumerateWithFiltersResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Return a list of backups for the specified volume",
        "tags": [
          "OpenStorageCloudBackup"
        ]
      }
    },
    "/v1/cloudbackups/history/{src_volume_id}": {
      "get": {
        "operationId": "History",
        "parameters": [
          {
            "in": "path",
            "name": "src_volume_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiSdkCloudBackupHistoryResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "History returns a list of backups for a specified volume",
        "tags": [
          "OpenStorageCloudBackup"
        ]
      }
    },
...
    "/v1/cloudbackups/schedules": {
      "get": {
        "operationId": "SchedEnumerate",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiSdkCloudBackupSchedEnumerateResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Enumerate cloud backup schedules",
        "tags": [
          "OpenStorageCloudBackup"
        ]
      },
      "post": {
        "operationId": "SchedCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/apiSdkCloudBackupSchedCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiSdkCloudBackupSchedCreateResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Create cloud backup schedule",
        "tags": [
          "OpenStorageCloudBackup"
        ]
      }
    },
...
    "/v1/cloudbackups/statechange": {
      "post": {
        "operationId": "StateChange",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/apiSdkCloudBackupStateChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiSdkCloudBackupStateChangeResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "StateChange can be used to stop, pause, and restart a backup",
        "tags": [
          "OpenStorageCloudBackup"
        ]
      }
    },
...
    "/v1/volumes/filters": {
      "post": {
        "operationId": "EnumerateWithFilters",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/apiSdkVolumeEnumerateWithFiltersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiSdkVolumeEnumerateWithFiltersResponse"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Enumerate returns a list of volume ids that match the labels if any are provided.",
        "tags": [
          "OpenStorageVolume"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}